requirements.txt
# --- API & Web Server ---
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1 

# --- Machine Learning & Data ---
scikit-learn==1.7.2
joblib==1.4.2
pandas==3.0.0
numpy==2.4.2